fix(service-portal): refactor signee view
