+++ libs/service-portal/signature-collection/src/screens/shared/SigneeView/index.tsx
@@ -2,14 +2,12 @@ import {
   ActionCard,
   AlertMessage,
   Box,
-  Button,
   Stack,
   Text,
 } from '@island.is/island-ui/core'
 import { useLocale } from '@island.is/localization'
 import { EmptyState } from '@island.is/service-portal/core'
 import { useGetListsForUser, useGetSignedList } from '../../../hooks'
-import format from 'date-fns/format'
 import { Skeleton } from '../../../skeletons'
 import { useAuth } from '@island.is/auth/react'
 import { sortAlpha } from '@island.is/shared/utils'
@@ -23,7 +21,6 @@ const SigneeView = ({
   currentCollection: SignatureCollection
 }) => {
   const { userInfo: user } = useAuth()
-
   const { formatMessage } = useLocale()
   const { signedLists, loadingSignedLists } = useGetSignedList()
   const { listsForUser, loadingUserLists } = useGetListsForUser(
@@ -46,7 +43,7 @@ const SigneeView = ({
             )}
           <Box marginTop={[2, 7]}>
             {/* Signed list */}
-            <SignedList />
+            <SignedList currentCollection={currentCollection} />
 
             {/* Other available lists */}
             <Box marginTop={[5, 10]}>
@@ -63,12 +60,12 @@ const SigneeView = ({
                       key={list.id}
                       backgroundColor="white"
                       heading={list.title}
-                      eyebrow={
-                        formatMessage(m.endTime) +
-                        ' ' +
-                        format(new Date(list.endTime), 'dd.MM.yyyy')
+                      eyebrow={' '}
+                      text={
+                        currentCollection.isPresidential
+                          ? formatMessage(m.collectionTitle)
+                          : formatMessage(m.collectionTitleParliamentary)
                       }
-                      text={formatMessage(m.collectionTitle)}
                       cta={
                         new Date(list.endTime) > new Date() && !list.maxReached
                           ? {
