+++ libs/service-portal/signature-collection/src/skeletons.tsx
@@ -3,7 +3,7 @@ import { Box, SkeletonLoader } from '@island.is/island-ui/core'
 export const Skeleton = () => {
   return (
     <Box marginTop={[5, 10]}>
-      <SkeletonLoader height={200} borderRadius="large" repeat={4} space={3} />
+      <SkeletonLoader height={150} borderRadius="large" repeat={3} space={2} />
     </Box>
   )
 }
